refactor: directory name
